fix: restore compile/test-green baseline for Slice 2 + collision-resolution work
The Slice 2 and schema-collision commits (7cf2abc, 2123990, f18710a,
c83e88c) did not actually compile, and the test target did not
compile either -- this was only discovered now because this is the
first time anyone ran `swift build`/`swift test` against them with a
real toolchain in this session. Prior status docs claiming "297
tests, all green" were not reflecting reality.

Compile fixes (no behavior change, restores the evidently-intended API):
- Row has .int(_:), never .int64(_:) -- MoodLogStore, SorenessLogStore,
  CaffeineLogStore, InjuryNoteStore, DrinkEntryStore, VitalsRecordStore,
  VitalsRecordHealthBridge all called the nonexistent method.
- Migration015 declared `func run` and no `name`, which doesn't satisfy
  the Migration protocol (`static func up`, `static var name`) every
  other migration in this file implements. Fixed to match.
- decodeJSON<T: Decodable>() with `?? []` and no type annotation let
  Swift infer T as [Any] instead of [String] (ProfileStore.sports,
  SorenessLogStore.bodyAreas) -- added explicit `: [String]`.
- Two Int64-returning row reads assigned into Int-typed fields
  (MoodLogEntry.score, SorenessLogEntry.overallScore) without narrowing.
- Tests/AlmanacCoreTests/TestDatabase.swift did not exist. Ten new
  Swift Testing files (Mood/Soreness/Injury/Vitals/Profile stores,
  Sleep/VitalsRecord HealthKit bridges, readiness-cycle linking) all
  called `TestDatabase()` as their fixture -- the test target could not
  compile without it. Added: in-memory DB with every migration applied.

Runtime/schema fixes (found once the suite could actually run):
- mood_log, soreness_log, vitals_record all declared `timezoneOffset
  TEXT NOT NULL`, but every store's Draft type declares it `Int?` and
  defaults to nil -- every insert without an explicit offset violated
  the constraint. Relaxed to nullable, matching the field's own
  documented optionality.
- ProfileStore's singleton row (id=1) was never seeded anywhere, so
  every updateX() was a silent no-op UPDATE against zero rows. Added
  ensureRowExists() (INSERT OR IGNORE) called before each write.
- CoreDailySchemaTests.testEveryMigrationAppliesInOrder still asserted
  Array(1...14); never bumped when migration 015 shipped.
- Three readiness-cycle-linking tests (Mood/Soreness store tests plus
  the dedicated linking-service suite) referenced a readinessCycleId
  FK to a readiness_cycle row that was never inserted -- seeded one.

Net effect, full suite (Swift 6.3.3, this session):
  before: build did not compile at all.
  after:  279/297 passing. Remaining 18 XCTest failures are pre-existing
  and distinct: Migration015 replaced sync_anchor's shape but
  SyncAnchorStore/HealthSyncService/BackupService were never updated to
  match (BackupServiceTests, HealthSyncTests, HydrationSyncTests,
  SyncAnchorStoreTests, TimelineTests). The remaining 8 Swift Testing
  issues are VitalsRecordHealthBridge/SleepEpisodeHealthBridge writing
  to column names (startTime/endTime/recordedAt/recordedTzOffset,
  deletedAt) that don't exist in migration 014's actual sleep_episode/
  vitals_record tables. Both are real, substantive HealthKit-sync bugs,
  left unfixed here as out of Slice 4 scope -- see the project status
  note for detail; they belong with the iMac/HealthKit wiring track.

+++ Sources/AlmanacCore/Hydration/DrinkEntryStore.swift
@@ -119,7 +119,7 @@ public struct DrinkEntryStore: @unchecked Sendable {
         
         // Fetch the last inserted rowid
         if let row = try db.query("SELECT last_insert_rowid() as id;").first,
-           let fetchedID = row.int64("id") {
+           let fetchedID = row.int("id") {
             id = fetchedID
         }
         return id
@@ -157,7 +157,7 @@ public struct DrinkEntryStore: @unchecked Sendable {
     // MARK: - Private
 
     private func rowToEntry(_ row: Row) -> DrinkEntry? {
-        guard let id = row.int64("id"),
+        guard let id = row.int("id"),
               let timestamp = row.string("timestamp").flatMap(iso8601ToDate),
               let logicalDay = row.string("logicalDay"),
               let drinkType = row.string("drinkType") else { return nil }
