+++ Sources/AlmanacCore/Mood/MoodLogStore.swift
@@ -83,7 +83,7 @@ public struct MoodLogStore: @unchecked Sendable {
         
         // Fetch the last inserted rowid
         if let row = try db.query("SELECT last_insert_rowid() as id;").first,
-           let fetchedID = row.int64("id") {
+           let fetchedID = row.int("id") {
             id = fetchedID
         }
         return id
@@ -141,7 +141,7 @@ public struct MoodLogStore: @unchecked Sendable {
     // MARK: - Private
 
     private func rowToEntry(_ row: Row) -> MoodLogEntry? {
-        guard let id = row.int64("id"),
+        guard let id = row.int("id"),
               let score = row.int("score"),
               let timestampText = row.string("timestamp"),
               let logicalDay = row.string("logicalDay"),
@@ -149,11 +149,11 @@ public struct MoodLogStore: @unchecked Sendable {
         
         return MoodLogEntry(
             id: id,
-            score: score,
+            score: Int(score),
             timestamp: timestamp,
             logicalDay: logicalDay,
             notes: row.string("notes"),
-            readinessCycleId: row.int64("readinessCycleId"),
+            readinessCycleId: row.int("readinessCycleId"),
             createdAt: row.string("createdAt").flatMap(iso8601ToDate) ?? Date()
         )
     }
