+++ Sources/AlmanacCore/Readiness/CaffeineLogStore.swift
@@ -108,7 +108,7 @@ public struct CaffeineLogStore: @unchecked Sendable {
         
         // Fetch the last inserted rowid
         if let row = try db.query("SELECT last_insert_rowid() as id;").first,
-           let fetchedID = row.int64("id") {
+           let fetchedID = row.int("id") {
             id = fetchedID
         }
         return id
@@ -156,7 +156,7 @@ public struct CaffeineLogStore: @unchecked Sendable {
     // MARK: - Private
 
     private func rowToEntry(_ row: Row) -> CaffeineLogEntry? {
-        guard let id = row.int64("id"),
+        guard let id = row.int("id"),
               let timestamp = row.string("timestamp").flatMap(iso8601ToDate),
               let logicalDay = row.string("logicalDay"),
               let drinkType = row.string("drinkType"),
@@ -168,7 +168,7 @@ public struct CaffeineLogStore: @unchecked Sendable {
             logicalDay: logicalDay,
             drinkType: drinkType,
             caffeineMg: caffeineMg,
-            drinkEntryId: row.int64("drinkEntryId"),
+            drinkEntryId: row.int("drinkEntryId"),
             intendedSleepTimestamp: row.string("intendedSleepTimestamp").flatMap(iso8601ToDate),
             hoursBeforeIntendedSleep: row.double("hoursBeforeIntendedSleep"),
             caffeineContext: row.string("caffeineContext"),
