+++ Sources/AlmanacCore/Vitals/VitalsRecordStore.swift
@@ -98,7 +98,7 @@ public struct VitalsRecordStore: @unchecked Sendable {
         
         // Fetch the last inserted rowid
         if let row = try db.query("SELECT last_insert_rowid() as id;").first,
-           let fetchedID = row.int64("id") {
+           let fetchedID = row.int("id") {
             id = fetchedID
         }
         return id
@@ -155,7 +155,7 @@ public struct VitalsRecordStore: @unchecked Sendable {
     // MARK: - Private
 
     private func rowToRecord(_ row: Row) -> VitalsRecord? {
-        guard let id = row.int64("id"),
+        guard let id = row.int("id"),
               let metric = row.string("metric"),
               let value = row.double("value"),
               let unit = row.string("unit"),
